+++ packages/@overeng/notion-cli/pnpm-workspace.yaml.genie.ts
@@ -4,6 +4,7 @@ import notionEffectClientPkg from '../notion-effect-client/package.json.genie.ts
 import notionEffectSchemaPkg from '../notion-effect-schema/package.json.genie.ts'
 import tuiCorePkg from '../tui-core/package.json.genie.ts'
 import tuiReactPkg from '../tui-react/package.json.genie.ts'
+import utilsDevPkg from '../utils-dev/package.json.genie.ts'
 import utilsPkg from '../utils/package.json.genie.ts'
 import pkg from './package.json.genie.ts'
 
@@ -15,6 +16,7 @@ export default pnpmWorkspaceWithDepsReact({
     notionEffectSchemaPkg,
     tuiCorePkg,
     tuiReactPkg,
+    utilsDevPkg,
     utilsPkg,
   ],
 })