+++ packages/@overeng/notion-effect-schema/package.json.genie.ts
@@ -25,7 +25,13 @@ export default packageJson({
     ...effectLspScripts,
   },
   devDependencies: {
-    ...catalog.pick(...peerDepNames, '@effect/vitest', '@types/node', 'vitest'),
+    ...catalog.pick(
+      ...peerDepNames,
+      '@effect/vitest',
+      '@overeng/utils-dev',
+      '@types/node',
+      'vitest',
+    ),
     ...effectLspDevDeps(),
   },
   peerDependencies: catalog.peers(...peerDepNames),