+++ packages/@overeng/tui-react/pnpm-workspace.yaml.genie.ts
@@ -2,4 +2,8 @@ import { pnpmWorkspaceWithDepsReact } from '../../../genie/internal.ts'
 import tuiCorePkg from '../tui-core/package.json.genie.ts'
 import pkg from './package.json.genie.ts'
 
-export default pnpmWorkspaceWithDepsReact({ pkg, deps: [tuiCorePkg] })
+export default pnpmWorkspaceWithDepsReact({
+  pkg,
+  deps: [tuiCorePkg],
+  extraPackages: ['../utils-dev'],
+})